Modified DTPHandler class which was implemented as a subclass of asyncore.dispatcher instead of asynchat.async_chat, following the same approach that asynchat module uses in Python 2.6, to work around issue #1740572 (http://bugs.python.org/issue1740572).
DTPHandler now inherits from asynchat.async_chat, overriding refill_buffer() method only (the buggy one).

This is good since we are no longer tied up to changes which might be applied to asynchat.py internals in future and also makes the porting to Python 3.X and Jython a lot easier.